+++ database/migrations/2022_06_09_155355_create_incomings_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->unsignedBigInteger('actor_id');
             $table->foreign('actor_id')->references('id')->on('actors')->onDelete('cascade');
             $table->json('products');
-            $table->float('total');
+            $table->decimal('total',10,2);
             $table->timestamps();
             $table->engine = 'InnoDB';
         });
